zonal: use int64 zone-boundary counter in _strides to avoid int32 overflow (#2612)
_strides accumulated cumulative zone-boundary counts into an int32 array
inside a numba @ngjit kernel. For rasters above ~2.1 billion pixels the
count wraps silently to a negative value, corrupting the slice bounds
that _calc_stats and crosstab derive from it, so stats() and crosstab()
returned silently wrong results on very large numpy-backed (or
large-chunk dask) rasters. Allocate the counter as int64 to match the
natural array-index type.

+++ xrspatial/zonal.py
@@ -284,7 +284,7 @@ def _parallel_variance(block_counts, block_sums, block_m2s):
 def _strides(flatten_zones, unique_zones):
     num_elements = flatten_zones.shape[0]
     num_zones = len(unique_zones)
-    strides = np.zeros(len(unique_zones), dtype=np.int32)
+    strides = np.zeros(len(unique_zones), dtype=np.int64)
 
     count = 0
     for i in range(num_zones):
